+++ changelog.md
@@ -33,6 +33,7 @@
 - Added class `RichText`
 - Added class `Reactions`
 - Added class `Emojis`
+- Added class `TrustedCertificates`
 - **[BREAKING]** Renamed method and class `GraphExecution` to `GraphExecutions`
 - **[BREAKING]** Renamed method and class `Job` to `Jobs`
 - **[BREAKING]** Renamed method and class `Repository` to `Repositories`